Update modulo.py
I changed the input and the output to integers

+++ modulo.py
@@ -1,2 +1,2 @@
-def modulo(x: float, y: float) -> float:
+def modulo(x: int, y: int) -> int:
     return x % y
